lots of int/float/double cleanup and servogui worky

+++ src/main/java/org/myrobotlab/service/BodyPart.java
@@ -268,25 +268,25 @@ public class BodyPart extends AbstractBodyPart implements IKJointAngleListener {
             if (angle < 0) {
               angle += 360;
             }
-            getServo("omoplate").moveTo(angle.intValue());
+            getServo("omoplate").moveTo(angle);
           }
           if ("shoulder".equals(s)) {
             Double angle = (gainMap.get(s) * angleMap.get(s) + phaseShiftMap.get(s)) % 360.0;
             if (angle < 0) {
               angle += 360;
             }
-            getServo("shoulder").moveTo(angle.intValue());
+            getServo("shoulder").moveTo(angle);
           }
           if ("rotate".equals(s)) {
             Double angle = (gainMap.get(s) * angleMap.get(s) + phaseShiftMap.get(s)) % 360.0;
             if (angle < 0) {
               angle += 360;
             }
-            getServo("rotate").moveTo(angle.intValue());
+            getServo("rotate").moveTo(angle);
           }
           if ("bicep".equals(s)) {
             Double angle = (gainMap.get(s) * angleMap.get(s) + phaseShiftMap.get(s)) % 360.0;
-            getServo("bicep").moveTo(angle.intValue());
+            getServo("bicep").moveTo(angle);
             if (angle < 0) {
               angle += 360;
             }
